[feat] Extract common layouts for options & Add OptionEntry
- 옵션 레이아웃 컴포넌트 OptionEntry 추가
- TextInputContainer -> TextInputOption 이름 변경
- TextInputOption OptionEntry 기반으로 재설계
- OptionContainer -> OptionBoxContainer 이름 변경 및 디렉토리 이동
